requirements.txt
aiohttp==3.8.3
authlib==1.2.0
fastapi==0.88.0
httpx==0.23.1
itsdangerous==2.1.2
python-jose==3.3.0
python-multipart==0.0.5
starlette==00.22.0
uvicorn==0.20.0